# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Agendamento API",
        "description": "API REST para gerenciamento de agendamentos com autenticação JWT, gestão de dias fechados e QR Code para validação. Esta API permite criar, listar, visualizar e cancelar agendamentos, além de gerenciar dias bloqueados e validar QR Codes.",
        "termsOfService": "http://localhost:8000/terms",
        "contact": {
            "name": "Equipe de Desenvolvimento",
            "url": "https://github.com/seu-usuario/agendamento",
            "email": "dev@agendamento.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login do usuário",
                "description": "Autentica um usuário e retorna um token de acesso",
                "operationId": "42f73b33a1e0a43645f398cf4816eb60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Usuário ou senha inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Redefinir senha",
                "description": "Redefine a senha do usuário usando token de reset",
                "operationId": "fd7ba52856eb1daa766b470920f1ccdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "nova_senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    },
                                    "nova_senha": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha redefinida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Senha redefinida com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na redefinição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Token inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/password/request-reset": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Solicitar reset de senha",
                "description": "Solicita um token para reset de senha por email",
                "operationId": "1d61d671c874272f26a06bf4ce728176",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token enviado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Token enviado para o e-mail se existir um usuário com esse e-mail."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/validar-qrcode/{uuid}": {
            "get": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Validar QR Code",
                "description": "Valida um QR Code de agendamento",
                "operationId": "ec59a5401563eef45a72398f00a95390",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "abc123-def456-ghi789"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR Code válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "QR Code válido"
                                        },
                                        "agendamento": {
                                            "properties": {
                                                "data": {
                                                    "type": "string",
                                                    "example": "15/12/2023"
                                                },
                                                "horario": {
                                                    "type": "string",
                                                    "example": "14:00"
                                                },
                                                "grupo": {
                                                    "type": "string",
                                                    "example": "Sim"
                                                },
                                                "quantidade": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "observacao": {
                                                    "type": "string",
                                                    "example": "Observação do agendamento"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "QR Code inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "QR Code inválido ou agendamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Digite o token Bearer no formato: Bearer {seu-token-jwt}"
            }
        }
    },
...
}